% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_sim.R
\name{acceptance_sample}
\alias{acceptance_sample}
\title{Acceptance Sample}
\usage{
acceptance_sample(
  n,
  pdf,
  c = 10,
  compare.dist = "norm",
  compare.params = c(0, 1)
)
}
\arguments{
\item{n}{sample size}

\item{pdf}{distribution to sample from}

\item{c}{accept/reject costant}

\item{compare.dist}{comparison/proposal distribution, one of c('norm','uniform')}

\item{compare.params}{parameters for compare.dist, either c(mean, sd) or c(min, max)}
}
\value{
vector of sampled values
}
\description{
Acceptance Sample
}
