% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_sim.R
\name{alternative}
\alias{alternative}
\title{Alternative pdf}
\usage{
alternative(
  t,
  sigma_1 = NULL,
  sigma_1L = NULL,
  sigma_1R = NULL,
  pi_1L = 0.5,
  type = "symmetric"
)
}
\arguments{
\item{t}{test statistic}

\item{sigma_1}{variance of alternative distribution if symmetric}

\item{sigma_1L}{variance of negative alternative distribution if asymmetric}

\item{sigma_1R}{variance of positive alternative distribution if asymmetric}

\item{pi_1L}{proportion of alternative distribution > 0 if asymmetric}

\item{type}{one of c('symmetric','asymmetric')}
}
\value{
value of alternative density pdf at value t
}
\description{
Alternative pdf
}
