% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_fdrtool_grid.R
\name{build_fdrtool_grid}
\alias{build_fdrtool_grid}
\title{Build fdrtool Grid}
\usage{
build_fdrtool_grid(
  t,
  cutoff.method = c("fndr", "pct0", "locfdr"),
  pct0 = 1/c(2:10),
  verbose = FALSE
)
}
\arguments{
\item{t}{vector of test statistics}

\item{cutoff.method}{vector of options for cutoff method hyperparameter}

\item{pct0}{vector of options for pct0 hyperparameter}
}
\value{
dataframe where each row is a possible set of hyperparameters for
fdrtool on a specific set of test statistics
}
\description{
Build a grid of possible hyperparameters for fdrtool
from separate vectors of hyperparameter options. Final grid only considers
hyperparameters that can be run on provided data without error.
}
