% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_locfdr_grid.R
\name{build_locfdr_grid}
\alias{build_locfdr_grid}
\title{Build locfdr Grid}
\usage{
build_locfdr_grid(
  test_statistics,
  pct_range = c(0, 0.1),
  pct0_range = c(0, 1/3),
  nulltype = 1:3,
  type = 0:1,
  grid_size = 40,
  method = "grid",
  lower_pi0 = 0.7,
  seed = NULL,
  parallel_param = NULL,
  verbose = FALSE
)
}
\arguments{
\item{test_statistics}{vector, test statistics}

\item{pct_range}{vector c(min, max), range for pct hyperparameter. \code{pct} is the
excluded tail proportions of zz's when fitting f(z).
pct=0 includes full range of zz's.}

\item{pct0_range}{vector c(min, max), range for pct0 hyperparameter. \code{pct0} is the
proportion of the zz distribution used in fitting the null density
f0(z) by central matching.}

\item{nulltype}{vector, options for nulltype hyperparameter. \code{nulltype}
is the type of null hypothesis assumed in estimating f0(z), for
use in the fdr calculations. 0 is the theoretical null N(0,1),
1 is maximum likelihood estimation, 2 is central matching estimation,
3 is a split normal version of 2.}

\item{type}{vector, options for type hyperparameter. \code{type} is the type of
fitting used for f; 0 is a natural spline, 1 is a polynomial.}

\item{parallel_param}{BiocParallel object, specified to run in parallel or NULL}

\item{verbose}{boolean}
}
\value{
data.frame, each row is a possible set of hyperparameters for locfdr
}
\description{
Build a factorial grid of possible hyperparameters for locfdr and
reduce to hyperparameter sets that can be run on provided data without error.
}
