% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_fdrtool_grid.R
\name{check_fdrtool_row}
\alias{check_fdrtool_row}
\title{Check Fdrtool Row}
\usage{
check_fdrtool_row(test_statistics, fdrtool_grid, row, lower_pi0)
}
\arguments{
\item{test_statistics}{vector, test statistics}

\item{fdrtool_grid}{data.frame, each row is a possible set of hyperparameters for fdrtool}

\item{row}{integer, row of fdrtool_grid considered}

\item{lower_pi0}{double, hyperparameter combinations that give pi0 estimates below this are excluded from the grid}
}
\value{
boolean, whether the hyperparameter combination run without error
}
\description{
For a given row of fdrtool_grid, checks whether hyperparameter combinations
can be used for fdrtool on the provided data without error
}
