% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_locfdr_grid.R
\name{check_locfdr_row}
\alias{check_locfdr_row}
\title{Check locfdr Row
For a given row of locfdr_grid, checks whether hyperparameter combinations
can be used for locfdr on the provided data without error}
\usage{
check_locfdr_row(test_statistics, locfdr_grid, row, lower_pi0)
}
\arguments{
\item{test_statistics}{vector, test statistics}

\item{locfdr_grid}{data.frame, each row is a possible set of hyperparameters for locfdr}

\item{row}{integer, row of locfdr_grid considered}

\item{lower_pi0}{double, exclude hyperparameter combinations that give pi0 estimates below this threshold}
}
\value{
boolean, whether the hyperparameter combination ran without error
}
\description{
Check locfdr Row
For a given row of locfdr_grid, checks whether hyperparameter combinations
can be used for locfdr on the provided data without error
}
