% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_qvalue_grid.R
\name{check_qvalue_row}
\alias{check_qvalue_row}
\title{Check qvalue Row}
\usage{
check_qvalue_row(test_statistics, qvalue_grid, lower_pi0, row, df)
}
\arguments{
\item{test_statistics}{vector, test statistics}

\item{qvalue_grid}{data.frame, each row is a possible set of hyperparameters for locfdr}

\item{lower_pi0}{double, exclude hyperparameter combinations that give pi0 estimates below this threshold}

\item{row}{integer, row of qvalue_grid considered}

\item{df}{integer, degrees of freedom of test statistics, if known}
}
\value{
boolean, whether the hyperparameter combination ran without error
}
\description{
For a given row of qvalue_grid, checks whether hyperparameter combinations
can be used for qvalue on the provided data without error
}
