% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{classification_metrics}
\alias{classification_metrics}
\title{Compute Classification Metrics}
\usage{
classification_metrics(fdr, test_statistics, truth, cutoff)
}
\arguments{
\item{fdr}{vector, local fdr values}

\item{test_statistics}{vector, test statistics}

\item{truth}{vector, truth values}

\item{cutoff}{double, fdr <  cutoff predicted as not-null}
}
\value{
named list of metrics based on specified cutoff
}
\description{
Compute Classification Metrics
}
