% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{ensemble}
\alias{ensemble}
\title{Ensemble}
\usage{
ensemble(
  test_statistics,
  top_grid,
  locfdr_grid,
  fdrtool_grid,
  qvalue_grid,
  df = NULL,
  focus_metric = "fdrerror",
  large_abs_metric = FALSE,
  parallel_param = NULL,
  verbose = TRUE
)
}
\arguments{
\item{test_statistics}{vector, test statistics}

\item{top_grid}{data.frame, specifies packages and row numbers within that grid of
the best methods used in the ensemble}

\item{locfdr_grid}{data.frame, each row is a set of hyperparameters for locfdr}

\item{fdrtool_grid}{data.frame, each row is a set of hyperparameters for fdrtool}

\item{qvalue_grid}{data.frame, each row is a set of hyperparameters for qvalue}

\item{df}{integer, degrees of freedom of test statistics, if known}

\item{focus_metric}{string, one of one of c('fdrerror'),
which metric to optimize in the grid search}

\item{large_abs_metric}{boolean, if TRUE, only consider focus_metric looking at the
large absolute value test statistics (top quartile of abs(t))}

\item{parallel_param}{BiocParallel object, specified to run in parallel or NULL}

\item{verbose}{boolean}
}
\value{
\itemize{
\item fdr - estimated local false discovery rates
\item pi0 - estimated proportion of tests that are null
}
}
\description{
Ensemble
}
