% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fdr_from_fdr.R
\name{Fdr_from_fdr}
\alias{Fdr_from_fdr}
\title{Fdr from fdr}
\usage{
Fdr_from_fdr(fdr, test_statistics, direction = "left")
}
\arguments{
\item{fdr}{vector, local false discovery rate estimates}

\item{test_statistics}{vector, test statistics}

\item{direction}{string, one of c('left','right'), direction of tail-end false discovery rate}
}
\value{
vector, tail-end false discovery rates
}
\description{
Calculate tail-end false discovery rate (Fdr)
from local false discovery rate (fdr)
}
