% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generating_model.R
\name{fit_generating_model}
\alias{fit_generating_model}
\title{Fit generating model using EM algorithm}
\usage{
fit_generating_model(
  test_statistics,
  sigmasq0 = 2,
  sigmasq1 = 4,
  pi0 = 0.9,
  maxiter = 500,
  tol = 1e-04
)
}
\arguments{
\item{test_statistics}{vector, test statistics}

\item{sigmasq0}{double, initial value for sigmasq0}

\item{sigmasq1}{double, initial value for sigmasq1}

\item{pi0}{double, initial value for pi0}

\item{maxiter}{integer, highest number of iterations of EM algorithm allowed}

\item{tol}{double, tolerance for change in sum of squared differences in parameters
in order to stop algorithm}
}
\value{
list, named parameters for the generating_model densities and values across iterations
}
\description{
Fit generating model using EM algorithm
}
