% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generating_model.R
\name{fit_generating_model_symmetric}
\alias{fit_generating_model_symmetric}
\title{Fit symmetric generating model using EM algorithm}
\usage{
fit_generating_model_symmetric(
  test_statistics,
  sigmasq0,
  sigmasq1,
  pi0,
  maxiter,
  tol
)
}
\arguments{
\item{test_statistics}{vector, test statistics}

\item{sigmasq0}{double, initial value for sigmasq0}

\item{sigmasq1}{double, initial value for sigmasq1}

\item{pi0}{double, initial value for pi0}

\item{maxiter}{integer, max number of iterations of EM algorithm allowed}

\item{tol}{double, tolerance for change in sum of squared differences in parameters
in order to stop algorithm}
}
\value{
list, named parameters for the generating_model densities and values across iterations
}
\description{
Fit symmetric generating model using EM algorithm
}
