% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_sim.R
\name{fit_sim}
\alias{fit_sim}
\title{Fit for Simulation}
\usage{
fit_sim(t, type = "symmetric")
}
\arguments{
\item{t}{vector of test statistics}

\item{type}{one of c('symmetric','asymmetric')}
}
\value{
vector of parameters for the fit densities
}
\description{
Fit a null and alternative distribution in order to
simulate labeled data for the grid search.
}
