% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_sim.R
\name{fit_sim_asymmetric}
\alias{fit_sim_asymmetric}
\title{Fit for Simulation using EM algorithm, Asymmetric}
\usage{
fit_sim_asymmetric(
  test_statistics,
  sigmasq0 = 2,
  sigmasq1n = 4,
  sigmasq1p = 4,
  pi1n = 0.5,
  pi0 = 0.9,
  maxiter = 500,
  tol = 1e-04
)
}
\arguments{
\item{test_statistics}{vector of test statistics}

\item{sigmasq0}{initial value for sigmasq0}

\item{sigmasq1n}{initial value for sigmasq1n}

\item{sigmasq1p}{initial value for sigmasq1p}

\item{pi1n}{initial value for pi1n}

\item{pi0}{initial value for pi0}

\item{maxiter}{highest number of iterations of EM algorithm allowed}

\item{tol}{tolerance for change in sum of squared differences in parameters
in order to stop algorithm}
}
\value{
vector of named parameters for the fit densities and values across iterations
}
\description{
Fit a null and alternative distribution in order to
simulate labeled data for the grid search.
}
