% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_sim.R
\name{fit_sim_symmetric}
\alias{fit_sim_symmetric}
\title{Fit Symmetric Working Model using EM algorithm}
\usage{
fit_sim_symmetric(
  test_statistics,
  sigmasq0 = 2,
  sigmasq1 = 4,
  pi0 = 0.9,
  maxiter = 500,
  tol = 1e-04
)
}
\arguments{
\item{test_statistics}{vector, test statistics}

\item{sigmasq0}{double, initial value for sigmasq0}

\item{sigmasq1}{double, initial value for sigmasq1}

\item{pi0}{double, initial value for pi0}

\item{maxiter}{integer, max number of iterations of EM algorithm allowed}

\item{tol}{double, tolerance for change in sum of squared differences in parameters
in order to stop algorithm}
}
\value{
list, named parameters for the fit densities and values across iterations
}
\description{
Fit a null and alternative distribution in order to
simulate labeled data for the grid search.
}
