% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generating_model.R
\name{fit_synthetic_generator}
\alias{fit_synthetic_generator}
\title{Fit synthetic generator using EM algorithm}
\usage{
fit_synthetic_generator(
  test_statistics,
  sigmasq0 = 1,
  sigmasq1 = 2,
  pi0 = 0.9,
  sigmasq0_fixed = FALSE,
  sigmasq1_fixed = FALSE,
  pi0_fixed = FALSE,
  maxiter = 500,
  tol = 1e-04,
  verbose = TRUE,
  type = "symmetric"
)
}
\arguments{
\item{test_statistics}{vector, test statistics}

\item{sigmasq0}{double, initial value for sigmasq0}

\item{sigmasq1}{double, initial value for sigmasq1}

\item{pi0}{double, initial value for pi0}

\item{sigmasq0_fixed}{logical, whether sigmasq0 should be learned or kept at sigmasq0}

\item{sigmasq1_fixed}{logical, whether sigmasq1 should be learned or kept at sigmasq1}

\item{pi0_fixed}{logical, whether pi0 should be learned or kept at pi0}

\item{maxiter}{integer, highest number of iterations of EM algorithm allowed}

\item{tol}{double, tolerance for change in sum of squared differences in
parameters in order to stop algorithm}

\item{type}{string, type of null distribution, one of c("Normal","t")}

\item{standardize}{logical, whether to divide by standard deviation before fitting}

\item{standardize_by}{}
}
\value{
list, named parameters for the synthetic_generator densities and values
across iterations
}
\description{
Fit synthetic generator using EM algorithm
}
