% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generating_model.R
\name{fit_working_model}
\alias{fit_working_model}
\title{Fit working model using EM algorithm}
\usage{
fit_working_model(
  test_statistics,
  df = NULL,
  type = "Normal",
  standardize = FALSE,
  sigmasq0 = 2,
  sigmasq1 = 4,
  pi0 = 0.9,
  sigmasq0_fixed = FALSE,
  sigmasq1_fixed = FALSE,
  pi0_fixed = FALSE,
  maxiter = 500,
  tol = 1e-04,
  verbose = TRUE
)
}
\arguments{
\item{test_statistics}{vector, test statistics}

\item{df}{integer, degrees of freedom, required if \code{type = "t"}}

\item{type}{string, type of null distribution, one of c("Normal","t")}

\item{standardize}{logical, whether to divide by standard deviation before fitting}

\item{sigmasq0}{double, initial value for sigmasq0}

\item{sigmasq1}{double, initial value for sigmasq1}

\item{pi0}{double, initial value for pi0}

\item{sigmasq0_fixed}{logical, whether sigmasq0 should be learned or kept at sigmasq0}

\item{sigmasq1_fixed}{logical, whether sigmasq1 should be learned or kept at sigmasq1}

\item{pi0_fixed}{logical, whether pi0 should be learned or kept at pi0}

\item{maxiter}{integer, highest number of iterations of EM algorithm allowed}

\item{tol}{double, tolerance for change in sum of squared differences in
parameters in order to stop algorithm}
}
\value{
list, named parameters for the working_model densities and values
across iterations
}
\description{
Fit working model using EM algorithm
}
