% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{get_brier}
\alias{get_brier}
\title{Calculate Brier Score}
\usage{
get_brier(fdr, truth)
}
\arguments{
\item{fdr}{estimated fdr values}

\item{truth}{truth values, vector of 0 and 1s}
}
\value{
Brier Score
}
\description{
Calculate Brier Score
}
