% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{get_Fdr_error}
\alias{get_Fdr_error}
\title{Calculate Fdr error}
\usage{
get_Fdr_error(Fdr, true_Fdr)
}
\arguments{
\item{Fdr}{estimated (tail end) Fdr values}

\item{true_Fdr}{true Fdr values}
}
\value{
Fdr error
}
\description{
Calculate Fdr error
}
