% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{get_MSE}
\alias{get_MSE}
\title{Calculate MSE}
\usage{
get_MSE(estimate, true)
}
\arguments{
\item{estimate}{vector, estimated values}

\item{true}{vector, true values}
}
\value{
double, MSE
}
\description{
Calculate MSE
}
