% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_sim.R
\name{get_neg_log_likelihood}
\alias{get_neg_log_likelihood}
\title{Negative log likelihood of mixture density}
\usage{
get_neg_log_likelihood(Z, type = "symmetric")
}
\arguments{
\item{type}{one of c('symmetric','asymmetric')}

\item{t}{vector of test statistics}
}
\value{
function of parameters vector that computes NLL for the data, t
}
\description{
Negative log likelihood of mixture density
}
