% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{get_prauc}
\alias{get_prauc}
\title{Calculate PR AUC}
\usage{
get_prauc(fdr, truth)
}
\arguments{
\item{fdr}{estimated fdr values}

\item{truth}{truth values, vector of 0 and 1s}
}
\value{
PR AUC
}
\description{
Calculate PR AUC
}
