% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{get_roc}
\alias{get_roc}
\title{Calculate ROC AUC}
\usage{
get_roc(fdr, truth)
}
\arguments{
\item{fdr}{vector, estimated fdr values}

\item{truth}{vector, truth values}
}
\value{
double, ROC AUC
}
\description{
Calculate ROC AUC
}
