% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_search.R
\name{grid_search}
\alias{grid_search}
\title{Grid Search}
\usage{
grid_search(
  n,
  nsim,
  ensemble_size,
  fit,
  method_list,
  row_list,
  df = NULL,
  locfdr_grid = NULL,
  fdrtool_grid = NULL,
  qvalue_grid = NULL,
  focus_metric = "pr",
  large_abs_metric = TRUE,
  params_type = "symmetric",
  parallel_param = NULL,
  sim_n = NULL,
  verbose = TRUE
)
}
\arguments{
\item{n}{integer, number of test statistics in each simulated dataset}

\item{nsim}{integer, number of datasets to simulate.
If 0, no datasets are simulated and model(s) are randomly selected.}

\item{ensemble_size}{integer, number of models chosen to ensemble over OR double,
proportion of grid size to ensemble over}

\item{fit}{list, result of fit_sim()}

\item{df}{integer, degrees of freedom of test statistics, if known, or NULL}

\item{locfdr_grid}{data.frame, each row is a set of hyperparameters for locfdr}

\item{fdrtool_grid}{data.frame, each row is a set of hyperparameters for fdrtool}

\item{qvalue_grid}{data.frame, each row is a set of hyperparameters for qvalue}

\item{focus_metric}{string, one of one of c('fdrerror','Fdrerror','pr','roc','brier'),
which metric to optimize in the grid search}

\item{large_abs_metric}{boolean, if TRUE, only consider focus_metric looking at the
large absolute value test statistics (top quartile of abs(t))}

\item{params_type}{string, type of simulation model fit, one of c('symmetric')}

\item{parallel_param}{BiocParallel object, specified to run in parallel or NULL}

\item{sim_n}{integer, size of simulated datasets, default number of test statistics}

\item{verbose}{boolean}
}
\value{
\itemize{
\item fit - list, result of fit_sim()
\item top_grid - data.frame, each row corresponds to a model
to ensemble over. Each row references a method (locfdr, fdrtool,
or qvalue) and row numbers in the respective grid
\item all_grids - data.frame, simulation metrics across all
simulations and all models.
}
}
\description{
simulate data and run grid search to determine which
models to ensemble over on the real dataset
}
