% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridsemblefdr.R
\name{gridsemble}
\alias{gridsemble}
\title{Gridsemble}
\usage{
gridsemble(
  test_statistics,
  nsim = 10,
  ensemble_size = 10,
  locfdr_grid = NULL,
  fdrtool_grid = NULL,
  qvalue_grid = NULL,
  df = NULL,
  methods = c("locfdr", "fdrtool", "qvalue"),
  focus_metric = "fdrerror",
  lower_pi0 = 0.7,
  n_workers = NULL,
  large_abs_metric = TRUE,
  parallel = TRUE,
  parallel_param = NULL,
  sim_n = NULL,
  verbose = TRUE
)
}
\arguments{
\item{test_statistics}{vector, test statistics}

\item{nsim}{integer, number of datasets to simulate and grid search over.
If 0, no datasets are simulated and model(s) are randomly selected.}

\item{ensemble_size}{integer, number of models chosen to ensemble over.}

\item{locfdr_grid}{data.frame, each row is a set of hyperparameters for locfdr}

\item{fdrtool_grid}{data.frame, each row is a set of hyperparameters for fdrtool}

\item{qvalue_grid}{data.frame, each row is a set of hyperparameters for qvalue}

\item{df}{integer, degrees of freedom of test statistics, if known. Otherwise assumed
to be from N(0, 1).}

\item{methods}{vector, methods to be used. Default c('locfdr','fdrtool','qvalue')}

\item{focus_metric}{string, one of one of c('fdrerror','Fdrerror','pr','roc','brier'),
which metric to optimize in the grid search}

\item{large_abs_metric}{boolean, if TRUE, only consider focus_metric looking at the
large absolute value test statistics (top quartile of abs(t))}

\item{parallel}{boolean}

\item{sim_n}{integer, size of simulated datasets, default number of test statistics}

\item{verbose}{boolean}

\item{paralllel_param}{BiocParallel object, specified to run in parallel or NULL}
}
\value{
\itemize{
\item fdr local false discovery rates
\item Fdr left tail false discovery rates
\item pi0 proportion of null test statistics
\item top_grid dataframe containing hyperparameter sets
that were ensembled over and their metrics on simulated daa
\item all_grids dataframe containing all hyperparameter sets considered
and their metrics on simulated data
\item fit parameters used for simulation step
}
}
\description{
use ensemble methods to estimate local (fdr) and tail-end (Fdr)
false discovery rates.
}
