% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metrics}
\alias{metrics}
\title{Compute all metrics}
\usage{
metrics(
  fdr,
  Fdr,
  test_statistics,
  truth,
  true_Fdr,
  true_fdr,
  topq,
  cutoff = 0.2
)
}
\arguments{
\item{fdr}{vector, estimated fdr values}

\item{Fdr}{vector, estimated (tail end) Fdr values}

\item{truth}{vector, truth values, vector of 0 and 1s}

\item{true_Fdr}{vector, true Fdr values}

\item{true_fdr}{vector, true fdr values}

\item{topq}{vector, booleans for whether each test statistic
is in the largest absolute quantile}
}
\value{
labeled list of metrics
}
\description{
Compute all metrics
}
