% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_sim.R
\name{metropolis_hastings_step}
\alias{metropolis_hastings_step}
\title{One Metropolis Hastings step to update t}
\usage{
metropolis_hastings_step(t, sigmasq1)
}
\arguments{
\item{t}{current t value}

\item{sigmasq1}{variance of alternative distribution}
}
\value{
t next t value
}
\description{
One Metropolis Hastings step to update t
}
