% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generating_model.R
\name{mix}
\alias{mix}
\title{Mixture pdf}
\usage{
mix(t, pi0, sigmasq0, sigmasq1)
}
\arguments{
\item{t}{double, test statistic}

\item{pi0}{double, proportion null}

\item{sigmasq0}{double, variance of null distribution}

\item{sigmasq1}{double, variance of alternative distribution}
}
\value{
value of mixture density pdf at value t
}
\description{
Mixture pdf
}
