% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_sim.R
\name{mixture}
\alias{mixture}
\title{Mixture density}
\usage{
mixture(
  z,
  sigma_0,
  sigma_1,
  sigma_1L,
  sigma_1R,
  pi_0,
  pi_1L,
  type = "symmetric"
)
}
\arguments{
\item{sigma_0}{variance of null distribution}

\item{sigma_1}{variance of alternative distribution if symmetric}

\item{sigma_1L}{variance of negative alternative distribution if asymmetric}

\item{sigma_1R}{variance of positive alternative distribution if asymmetric}

\item{pi_1L}{proportion of alternative distribution > 0 if asymmetric}

\item{type}{one of c('symmetric','asymmetric')}

\item{t}{test statistic}
}
\value{
value of mixture density pdf at value t
}
\description{
Mixture density
}
