% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_search.R
\name{nrow_null0}
\alias{nrow_null0}
\title{Number of rows, 0 if null}
\usage{
nrow_null0(dataframe)
}
\arguments{
\item{dataframe}{data.frame or NULL}
}
\value{
integer, number of rows in the dataframe, or 0 if a
NULL value is passed in.
}
\description{
Number of rows, 0 if null
}
