% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_search.R
\name{oracle_grid_search}
\alias{oracle_grid_search}
\title{Oracle Grid Search}
\usage{
oracle_grid_search(
  test_statistics,
  true_Fdr,
  true_fdr = NULL,
  df = NULL,
  to_pval_function = function(test_statistics) {
     p_from_t(test_statistics, df = df)

    },
  locfdr_grid = "default",
  fdrtool_grid = "default",
  qvalue_grid = "default",
  focus_metric = "fdrerror",
  parallel = min(TRUE, n_workers > 1),
  n_workers = max(parallel::detectCores() - 2, 1),
  parallel_param = NULL,
  drop_pi0_1 = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{df}{integer, degrees of freedom of test statistics, if known, or NULL}

\item{to_pval_function}{function, converts test statistics vector to a
p-value vector. Default assumes t-distribution with given df under the null.}

\item{locfdr_grid}{data.frame, rows are possible hyperparameters for locfdr}

\item{fdrtool_grid}{data.frame, rows are possible hyperparameters for fdrtool}

\item{qvalue_grid}{data.frame, rows are possible hyperparameters for qvalue}

\item{focus_metric}{string, one of one of c('fdrerror'),
which metric to optimize in the grid search}

\item{parallel}{boolean, whether to utilize parallelization}

\item{n_workers}{integer, number of cores to use if parallel}

\item{parallel_param}{\code{BiocParallel} object}

\item{verbose}{boolean}

\item{synthetic_generator}{list, result of fit_synthetic_generator()}

\item{nsim}{integer, number of datasets to simulate.
If 0, no datasets are simulated and model(s) are randomly selected.}

\item{synthetic_size}{integer, sample size of each synthetic dataset}

\item{ensemble_size}{integer, number of models chosen to ensemble over}
}
\value{
\itemize{
\item top_grid - data.frame, each row corresponds to a model
to ensemble over. Each row references a method (locfdr, fdrtool,
or qvalue) and row numbers in the respective grid
\item all_grids - data.frame, simulation metrics across all
simulations and all models.
}
}
\description{
if the user has access to ground truth, this function
computes true fdr and Fdr MSE for each model in the grid
}
