% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{p_from_t}
\alias{p_from_t}
\title{P-value from t-statistic}
\usage{
p_from_t(test_statistics, df = NULL)
}
\arguments{
\item{test_statistics}{vector, test statistics}

\item{df}{integer, degrees of freedom to compute p-value from test statistics,
assume standard normal if NULL}
}
\value{
vector of 2-sided p-values
}
\description{
Get one-sided p-value from test statistic
}
