% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_parallel.R
\name{parlapply}
\alias{parlapply}
\title{Title}
\usage{
parlapply(X, FUN, parallel_param, ...)
}
\arguments{
\item{X}{a vector (atomic or list) or an expression object.}

\item{FUN}{the function to be applied to each element of X}

\item{parallel}{whether to run in parallel with \code{bplapply} or regularly with \code{lapply}}
}
\value{
result of lapply
}
\description{
Title
}
