% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generating_model.R
\name{proposal_pdf}
\alias{proposal_pdf}
\title{Proposal pdf}
\usage{
proposal_pdf(of, given, sd)
}
\arguments{
\item{of}{double, proposed value}

\item{given}{double, previous value}

\item{sd}{double, standard deviation of proposal distribution}
}
\value{
double, proposal pdf value
}
\description{
Proposal pdf
}
