% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_fdrtool_grid.R
\name{reduce_fdrtool_grid}
\alias{reduce_fdrtool_grid}
\title{Reduce fdrtool grid}
\usage{
reduce_fdrtool_grid(
  test_statistics,
  fdrtool_grid,
  parallel_param = NULL,
  verbose = FALSE
)
}
\arguments{
\item{test_statistics}{vector, test statistics}

\item{fdrtool_grid}{data.frame, rows are possible hyperparameters for fdrtool}

\item{parallel_param}{BiocParallel object}

\item{verbose}{boolean}
}
\value{
reduced fdrtool_grid: dataframe, rows are hyperparameters for fdrtool
}
\description{
Reduces grid to parameter combinations
that can run fdrtool on the provided data without error
}
