% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_locfdr_grid.R
\name{reduce_locfdr_grid}
\alias{reduce_locfdr_grid}
\title{Reduce locfdr grid}
\usage{
reduce_locfdr_grid(
  test_statistics,
  locfdr_grid,
  drop_pi0_1 = TRUE,
  parallel_param = NULL,
  verbose = FALSE
)
}
\arguments{
\item{test_statistics}{vector, test statistics}

\item{locfdr_grid}{data.frame, rows are possible hyperparameters for locfdr}

\item{drop_pi0_1}{boolean, whether to discard models that estimate pi0 = 1}

\item{parallel_param}{BiocParallel object}

\item{verbose}{boolean}
}
\value{
data.frame, each row is a possible set of hyperparameters for locfdr
}
\description{
Reduces grid to parameter combinations
that can run locfdr on the provided data without error
}
