% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_qvalue_grid.R
\name{reduce_qvalue_grid}
\alias{reduce_qvalue_grid}
\title{Reduce qvalue grid}
\usage{
reduce_qvalue_grid(
  test_statistics,
  to_pval_function,
  qvalue_grid,
  parallel_param = NULL,
  verbose = FALSE
)
}
\arguments{
\item{test_statistics}{vector, test statistics}

\item{to_pval_function}{function, converts test statistics vector to a
p-value vector.}

\item{qvalue_grid}{data.frame, rows are possible hyperparameters for qvalue}

\item{parallel_param}{BiocParallel object}

\item{verbose}{boolean}
}
\value{
data.frame, each row is a possible set of hyperparameters for qvalue
}
\description{
Reduces grid to parameter combinations
that can run qvalue on the provided data without error
}
