% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_locfdr_row.R
\name{run_locfdr_row}
\alias{run_locfdr_row}
\title{Run locfdr}
\usage{
run_locfdr_row(t, locfdr_grid, row)
}
\arguments{
\item{t}{vector of test statistics}

\item{locfdr_grid}{data frame where each row is a set of hyperparameters}

\item{row}{row of locfdr_grid, i.e. which set of hyperparameters to run locfdr with}
}
\value{

}
\description{
Run locfdr with a specific set of parameters
}
\examples{

}
