% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_qvalue_row.R
\name{run_qvalue_row}
\alias{run_qvalue_row}
\title{Run qvalue}
\usage{
run_qvalue_row(t, qvalue_grid, row, df = NULL, verbose = FALSE)
}
\arguments{
\item{t}{vector of test statistics}

\item{qvalue_grid}{data frame where each row is a set of hyperparameters}

\item{row}{row of qvalue_grid, i.e. which set of hyperparameters to run qvalue with}

\item{df}{degrees of freedom to compute p-value from test statistics,
assume standard normal if NULL}

\item{verbose}{}
}
\value{
\itemize{
\item fdr - estimated local false discovery rates
\item Fdr - estimated left tail-end false discovery rates
\item pi0 - estimated proportion of tests that are null
}
}
\description{
Run fdrtool with a specific set of parameters
}
