% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_row.R
\name{run_row}
\alias{run_row}
\title{Run grid row}
\usage{
run_row(
  test_statistics,
  df,
  grids,
  method,
  row,
  returnFdr = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{test_statistics}{vector, test statistics}

\item{df}{integer, degrees of freedom of test statistics, if known}

\item{grids}{list, grids named by method}

\item{method}{string, one of c('locfdr','fdrtool','qvalue')}

\item{row}{integer, which row of method grid to consider}

\item{returnFdr}{boolean, whether to calculate Fdr form fdr}

\item{verbose}{boolean}
}
\value{
\itemize{
\item fdr - estimated local false discovery rates
\item Fdr - estimated left tail-end false discovery rates if returnFdr = TRUE
\item pi0 - estimated proportion of tests that are null
}
}
\description{
Run grid row
}
