% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_sim.R
\name{sample_alternative}
\alias{sample_alternative}
\title{Sample from alternative density}
\usage{
sample_alternative(
  n,
  sigma_1 = NULL,
  sigma_1L = NULL,
  sigma_1R = NULL,
  pi_1L = 0.5,
  type = "symmetric"
)
}
\arguments{
\item{n}{sample size}

\item{sigma_1}{variance of alternative distribution if symmetric}

\item{sigma_1L}{variance of negative alternative distribution if asymmetric}

\item{sigma_1R}{variance of positive alternative distribution if asymmetric}

\item{pi_1L}{proportion of alternative distribution > 0 if asymmetric}

\item{type}{one of c('symmetric','asymmetric')}
}
\value{
vector of test statistics sampled from alternative density
}
\description{
Using acceptance sampling
}
