% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_sim.R
\name{simulate_from_fit}
\alias{simulate_from_fit}
\title{Simulate from fit}
\usage{
simulate_from_fit(n, fit, type = "symmetric")
}
\arguments{
\item{n}{sample size}

\item{fit}{result of fit_sim()}

\item{type}{one of c('symmetric','asymmetric')}
}
\value{
data frame of n simulated t-statistic truth pairs
}
\description{
Simulate from fit
}
