% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generating_model.R
\name{simulate_from_generating_model}
\alias{simulate_from_generating_model}
\title{Simulate from generating model}
\usage{
simulate_from_generating_model(n, generating_model, df = NULL)
}
\arguments{
\item{n}{integer, sample size}

\item{generating_model}{list, result of fit_generating_model()}

\item{df}{integer, degrees of freedom of test statistics, if known}
}
\value{
data.frame, n simulated t-statistic and truth pairs
}
\description{
Simulate from generating model
}
