% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generating_model.R
\name{simulate_from_synthetic_generator}
\alias{simulate_from_synthetic_generator}
\title{Simulate from synthetic generator}
\usage{
simulate_from_synthetic_generator(
  n,
  synthetic_generator,
  to_pval_function = p_from_t
)
}
\arguments{
\item{n}{integer, sample size}

\item{synthetic_generator}{list, result of fit_synthetic_generator()}

\item{to_pval_function}{function, converts test statistics vector to a
p-value vector.}
}
\value{
data.frame, n simulated t-statistic and truth pairs
}
\description{
Simulate from synthetic generator
}
