% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_search.R
\name{single_grid_search}
\alias{single_grid_search}
\title{Single Grid Search}
\usage{
single_grid_search(
  this_dat,
  df,
  method_list,
  row_list,
  fdrtool_grid,
  locfdr_grid,
  qvalue_grid,
  verbose
)
}
\arguments{
\item{this_dat}{list, result of simulate_from_generating_model()}

\item{df}{integer, degrees of freedom of test statistics, if known, or NULL}

\item{method_list}{vector, methods to consider}

\item{row_list}{vector, rows of each method grid to consider}

\item{fdrtool_grid}{data.frame, each row is a set of hyperparameters for fdrtool}

\item{locfdr_grid}{data.frame, each row is a set of hyperparameters for locfdr}

\item{qvalue_grid}{data.frame, each row is a set of hyperparameters for qvalue}

\item{verbose}{boolean}
}
\value{
data.frame where each row gives metrics for a model
}
\description{
run grid search on a given dataset
}
