% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_sim.R
\name{acceptance_sample}
\alias{acceptance_sample}
\title{Acceptance sample from a pdf}
\usage{
acceptance_sample(
  n,
  pdf,
  c = 10,
  compare_dist = "norm",
  compare_params = c(0, 1)
)
}
\arguments{
\item{n}{sample size}

\item{pdf}{distribution to sample from}

\item{c}{accept/reject constant}

\item{compare_dist}{comparison/proposal distribution, one of c('norm','uniform')}

\item{compare_params}{parameters for compare_dist, either c(mean, sd) or c(min, max)}
}
\value{
vector of sampled values
}
\description{
Acceptance sample from a pdf
}
