% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_sim.R
\name{alt}
\alias{alt}
\title{Alternative pdf}
\usage{
alt(t, sigmasq1)
}
\arguments{
\item{t}{double, test statistic}

\item{sigmasq1}{double, variance of alternative distribution if symmetric}
}
\value{
double, value of alternative density pdf at value t
}
\description{
Alternative pdf
}
