% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_fdrtool_grid.R
\name{build_fdrtool_grid}
\alias{build_fdrtool_grid}
\title{Build fdrtool Grid}
\usage{
build_fdrtool_grid(
  test_statistics,
  cutoff.method = c("fndr", "pct0", "locfdr"),
  pct0_range = c(0, 0.5),
  grid_size = 40,
  lower_pi0 = 0.7,
  method = "grid",
  seed = NULL,
  parallel = TRUE,
  n_workers = parallel::detectCores() - 2,
  parallel_param = NULL,
  verbose = FALSE
)
}
\arguments{
\item{test_statistics}{vector, test statistics}

\item{cutoff.method}{vector, options for cutoff method hyperparameter.
\code{cutoff.method} is one of "fndr" (default), "pct0", "locfdr".}

\item{pct0_range}{vector c(min, max), range for pct0 hyperparameter. \code{pct0} is the
fraction of data used for fitting null model - only if cutoff.method="pct0".}

\item{grid_size}{integer, maximum size of grid to use. Note that this is \emph{not the final grid size}
as some hyperparameter combinations may fail when run on the data.}

\item{lower_pi0}{double, exclude hyperparameter combinations that give pi0 estimates below this threshold}

\item{method}{string, one of c('random', 'grid'). 'random' will sample from a uniform
distribution within the ranges, 'grid' will select equally spaced values.}

\item{seed}{integer, random seed used if method = 'random'}

\item{parallel}{boolean, whether to utilize parallelization}

\item{n_workers}{integer, number of cores to use if parallel, default 2 less than available.}

\item{parallel_param}{BiocParallel object, specified to run in parallel or NULL}

\item{verbose}{boolean}
}
\value{
data.frame, each row is a possible set of hyperparameters for fdrtool
}
\description{
Build a factorial grid of possible hyperparameters for fdrtool and
reduce to hyperparameter sets that can be run on provided data without error.
}
