% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_fdrtool_grid.R
\name{build_fdrtool_grid}
\alias{build_fdrtool_grid}
\title{Build fdrtool Grid}
\usage{
build_fdrtool_grid(
  test_statistics,
  cutoff.method = c("fndr", "pct0", "locfdr"),
  pct0_range = c(0.4, 1),
  grid_depth = 5,
  parallel = min(TRUE, n_workers > 1),
  n_workers = max(parallel::detectCores() - 2, 1),
  parallel_param = NULL,
  verbose = FALSE
)
}
\arguments{
\item{test_statistics}{vector, test statistics}

\item{cutoff.method}{vector, options for cutoff method hyperparameter.
\code{cutoff.method} is one of "fndr" (default), "pct0", "locfdr".}

\item{pct0_range}{vector c(min, max), range of values for pct0 hyperparameter, the
fraction of data used for fitting null model - only if cutoff.method="pct0".
Default in \code{fdrtool} package is \code{pct0 = 0.75}.}

\item{grid_depth}{integer, number of evenly-spaced values of continuous parameters
considered within their respective \verb{_range}.}

\item{parallel}{boolean, whether to utilize parallelization}

\item{n_workers}{integer, number of cores to use if parallel}

\item{parallel_param}{BiocParallel object}

\item{verbose}{boolean}
}
\value{
data.frame, each row is a possible set of hyperparameters for fdrtool
}
\description{
Build a grid of possible hyperparameters for fdrtool and reduce
to hyperparameter sets that can be run on provided data without error.
}
\examples{
set.seed(123)
test_statistics = c(rnorm(800), runif(100, -10, -5), runif(100, 5, 10))
fdrtool_grid = build_fdrtool_grid(test_statistics)
}
