% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_locfdr_grid.R
\name{build_locfdr_grid}
\alias{build_locfdr_grid}
\title{Build locfdr Grid}
\usage{
build_locfdr_grid(
  t,
  pct = c(0, 2, 4, 6)/1000,
  pct0 = 1/c(3:10),
  nulltype = 1:3,
  type = 0:1,
  verbose = FALSE
)
}
\arguments{
\item{t}{vector of test statistics}

\item{pct}{vector of options for pct hyperparameter}

\item{pct0}{vector of options for pct0 hyperparameter}

\item{nulltype}{vector of options for nulltype hyperparameter}

\item{type}{vector of options for type hyperparameter}
}
\value{
dataframe where each row is a possible set of hyperparameters for
locfdr on a specific set of test statistics
}
\description{
Build a grid of possible hyperparameters for locfdr
from separate vectors of hyperparameter options. Final grid only considers
hyperparameters that can be run on provided data without error.
}
