% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_qvalue_grid.R
\name{build_qvalue_grid}
\alias{build_qvalue_grid}
\title{Build qvalue Grid}
\usage{
build_qvalue_grid(
  t,
  trasf = c("probit", "logit"),
  adj = c(0.8, 0.9, 1, 1.1, 1.2),
  pi0.method = c("bootstrap", "smoother"),
  smooth.log.pi0 = c(TRUE, FALSE),
  verbose = FALSE
)
}
\arguments{
\item{t}{vector of test statistics}

\item{trasf}{vector of options for trasf hyperparameter}

\item{adj}{vector of options for adj hyperparameter}

\item{pi0.method}{vector of options for pi0.method hyperparameter}

\item{smooth.log.pi0}{vector of options for smooth.log.pi0 hyperparameter}

\item{verbose}{}
}
\value{
dataframe where each row is a possible set of hyperparameters for
qvalue on a specific set of test statistics
}
\description{
Build a grid of possible hyperparameters for qvalue
from separate vectors of hyperparameter options. Final grid only considers
hyperparameters that can be run on provided data without error.
}
