% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_qvalue_grid.R
\name{build_qvalue_grid}
\alias{build_qvalue_grid}
\title{Build qvalue Grid}
\usage{
build_qvalue_grid(
  test_statistics,
  transf = c("probit", "logit"),
  adj = c(0.8, 0.9, 1, 1.1, 1.2),
  pi0.method = c("bootstrap", "smoother"),
  smooth.log.pi0 = c(TRUE, FALSE),
  df = NULL,
  parallel = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{test_statistics}{vector of test statistics}

\item{transf}{vector of options for transf hyperparameter. \code{transf} is
a transformation is applied to the p-values so that a local FDR estimate can
be formed that does not involve edge effects of the \link{0,1} interval in which
the p-values lie, either "probit" or "logit".}

\item{adj}{vector of options for adj hyperparameter. \code{adj} is a numeric value
that is applied as a multiple of the smoothing bandwidth used in the density
estimation.}

\item{pi0.method}{vector of options for pi0.method hyperparameter. \code{pi0.method}
is the method for automatically choosing tuning parameter in the estimation of
pi_0, the proportion of true null hypotheses, either "smoother" or "bootstrap".}

\item{smooth.log.pi0}{vector of options for smooth.log.pi0 hyperparameter.
If \code{smooth.log.pi0} is TRUE and pi0.method = "smoother", pi_0 will be estimated
by applying a smoother to a scatterplot of log(pi_0) estimates against the
tuning parameter lambda.}

\item{parallel}{if TRUE, processes are run in parallel}

\item{verbose}{if TRUE, status updates will be displayed}
}
\value{
dataframe where each row is a possible set of hyperparameters for
qvalue on a specific set of test statistics
}
\description{
Build a grid of possible hyperparameters for qvalue
from separate vectors of hyperparameter options. Final grid only considers
hyperparameters that can be run on provided data without error.
}
