% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_qvalue_grid.R
\name{check_qvalue_row}
\alias{check_qvalue_row}
\title{Check qvalue Row
For a given row of qvalue_grid, checks whether hyperparameter combinations
can be used for qvalue on the provided data without error}
\usage{
check_qvalue_row(test_statistics, qvalue_grid, lower_pi0, row)
}
\arguments{
\item{test_statistics}{vector, test statistics}

\item{row}{integer, row of locfdr_grid considered}

\item{locfdr_grid}{data.frame, each row is a possible set of hyperparameters for locfdr}
}
\value{
boolean, whether the hyperparameter combination ran without error
}
\description{
Check qvalue Row
For a given row of qvalue_grid, checks whether hyperparameter combinations
can be used for qvalue on the provided data without error
}
