% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{ensemble}
\alias{ensemble}
\title{Ensemble}
\usage{
ensemble(
  test_statistics,
  top_grid,
  locfdr_grid,
  fdrtool_grid,
  qvalue_grid,
  df = NULL,
  parallel = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{test_statistics}{vector of test statistics}

\item{top_grid}{data frame of method, row number combinations to ensemble}

\item{locfdr_grid}{data frame where each row is a set of hyperparameters for locfdr}

\item{fdrtool_grid}{data frame where each row is a set of hyperparameters for fdrtool}

\item{qvalue_grid}{data frame where each row is a set of hyperparameters for qvalue}

\item{df}{degrees of freedom of test statistics, if known}

\item{verbose}{}
}
\value{
\itemize{
\item fdr - estimated local false discovery rates
\item pi0 - estimated proportion of tests that are null
}
}
\description{
Ensemble
}
