% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fdr_from_fdr.R
\name{Fdr_from_fdr}
\alias{Fdr_from_fdr}
\title{Fdr from fdr}
\usage{
Fdr_from_fdr(fdr, t, direction = "left")
}
\arguments{
\item{fdr}{vector of local false discovery rate estimates}

\item{t}{vector of test statistics}

\item{direction}{direction of tail-end false discovery rate, c('left','right')}
}
\value{
vector of tail-end false discovery rates
}
\description{
Calculate tail-end false discovery rate (Fdr)
from local false discovery rate (fdr)
}
\examples{

}
