% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_sim.R
\name{fit_sim}
\alias{fit_sim}
\title{Fit working model using EM algorithm}
\usage{
fit_sim(
  test_statistics,
  type = "symmetric",
  sigmasq0 = 2,
  sigmasq1 = 4,
  sigmasq1n = 4,
  sigmasq1p = 4,
  pi1n = 0.5,
  pi0 = 0.9,
  maxiter = 500,
  tol = 1e-04
)
}
\arguments{
\item{test_statistics}{vector, test statistics}

\item{type}{string, one of c('symmetric'), set up this way to add other working model options later}

\item{sigmasq0}{double, initial value for sigmasq0}

\item{sigmasq1}{double, initial value for sigmasq1}

\item{pi0}{double, initial value for pi0}

\item{maxiter}{integer, highest number of iterations of EM algorithm allowed}

\item{tol}{double, tolerance for change in sum of squared differences in parameters
in order to stop algorithm}
}
\value{
vector of named parameters for the fit densities and values across iterations
}
\description{
Fit a null and alternative distribution in order to
simulate labeled data for the grid search.
}
