% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_search.R
\name{get_true_Fdr}
\alias{get_true_Fdr}
\title{Get true tail end Fdr from test statistics and truth labels}
\usage{
get_true_Fdr(test_statistics, truth)
}
\arguments{
\item{test_statistics}{vector, test statistics}

\item{truth}{vector, truth values}
}
\value{
vector, tail end Fdr values
}
\description{
Get true tail end Fdr from test statistics and truth labels
}
