% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_search.R
\name{grid_search}
\alias{grid_search}
\title{Grid Search}
\usage{
grid_search(
  generating_model,
  nsim,
  sim_size,
  ensemble_size,
  df = NULL,
  locfdr_grid = NULL,
  fdrtool_grid = NULL,
  qvalue_grid = NULL,
  focus_metric = "fdrerror",
  large_abs_metric = FALSE,
  parallel = TRUE,
  n_workers = parallel::detectCores() - 2,
  parallel_param = NULL,
  verbose = TRUE
)
}
\arguments{
\item{generating_model}{list, result of fit_generating_model()}

\item{nsim}{integer, number of datasets to simulate.
If 0, no datasets are simulated and model(s) are randomly selected.}

\item{sim_size}{integer, sample size of each simulation.}

\item{ensemble_size}{integer, number of models chosen to ensemble over OR double,
proportion of grid size to ensemble over}

\item{df}{integer, degrees of freedom of test statistics, if known, or NULL}

\item{locfdr_grid}{data.frame, each row is a set of hyperparameters for locfdr}

\item{fdrtool_grid}{data.frame, each row is a set of hyperparameters for fdrtool}

\item{qvalue_grid}{data.frame, each row is a set of hyperparameters for qvalue}

\item{focus_metric}{string, one of one of c('fdrerror'),
which metric to optimize in the grid search}

\item{large_abs_metric}{boolean, if TRUE, only consider focus_metric looking at the
large absolute value test statistics (top quartile of abs(t))}

\item{parallel}{boolean, whether to utilize parallelization}

\item{n_workers}{integer, number of cores to use if parallel, default 2 less than available.}

\item{parallel_param}{BiocParallel object, specified to run in parallel or NULL}

\item{verbose}{boolean}
}
\value{
\itemize{
\item generating_model - list, result of generating_model_sim()
\item top_grid - data.frame, each row corresponds to a model
to ensemble over. Each row references a method (locfdr, fdrtool,
or qvalue) and row numbers in the respective grid
\item all_grids - data.frame, simulation metrics across all
simulations and all models.
}
}
\description{
simulate data and run grid search to determine which
models to ensemble over on the real dataset
}
