% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_search.R
\name{grid_search}
\alias{grid_search}
\title{Grid Search}
\usage{
grid_search(
  n,
  nsim,
  topn,
  fit,
  method_list,
  row_list,
  df = NULL,
  locfdr_grid = NULL,
  fdrtool_grid = NULL,
  qvalue_grid = NULL,
  focus_metric = "pr",
  large_abs_metric = TRUE,
  params_type = "symmetric",
  parallel = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{n}{number of test statistics in the real dataset = number of
test statistics to simulate}

\item{nsim}{Number of datasets to simulate and grid search over.
If 0, no datasets are simulated and model(s) are randomly selected.}

\item{topn}{Number of models chosen from each simulation to ensemble over.
If \code{nsim = 0}, number of random models to ensemble over.}

\item{fit}{result of fit_sim()}

\item{df}{degrees of freedom of test statistics, if known}

\item{locfdr_grid}{data frame where each row is a set of hyperparameters for locfdr}

\item{fdrtool_grid}{data frame where each row is a set of hyperparameters for fdrtool}

\item{qvalue_grid}{data frame where each row is a set of hyperparameters for qvalue}

\item{focus_metric}{which metric to prioritize in the grid search.
Must be one of c('pr','roc','brier','Fdrerror')}

\item{large_abs_metric}{if TRUE, only consider focus_metric looking at the
large absolute value test statistics (specifically, top quartile of abs(t))}

\item{params_type}{type of simulation model fit, one of c('symmetric','asymmetric')}

\item{parallel}{if TRUE, process is run in parallel}

\item{verbose}{}
}
\value{
\itemize{
\item fit - result of fit_sim()
\item top_grid - data frame where each row corresponds to a model
to ensemble over. Each row references a method (locfdr, fdrtool,
or qvalue) and row numbers in the respective grid
\item all_grids - data frame with simulation metrics across all
simulations and all models.
}
}
\description{
simulate data and run grid search(es) to determine which
models to ensemble over on the real dataset
}
