% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridsemblefdr.R
\name{gridsemble}
\alias{gridsemble}
\title{Gridsemble}
\usage{
gridsemble(
  test_statistics,
  nsim = 10,
  topn = 1,
  locfdr_grid = NULL,
  fdrtool_grid = NULL,
  qvalue_grid = NULL,
  df = NULL,
  methods = c("locfdr", "fdrtool", "qvalue"),
  asym = FALSE,
  focus_metric = "pr",
  large_abs_metric = TRUE,
  parallel = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{test_statistics}{vector of test statistics}

\item{nsim}{Number of datasets to simulate and grid search over.
If 0, no datasets are simulated and model(s) are randomly selected.}

\item{topn}{Number of models chosen from each simulation to ensemble over.
If \code{nsim = 0}, number of random models to ensemble over.}

\item{locfdr_grid}{data frame where each row is a set of hyperparameters for locfdr}

\item{fdrtool_grid}{data frame where each row is a set of hyperparameters for fdrtool}

\item{qvalue_grid}{data frame where each row is a set of hyperparameters for qvalue}

\item{df}{degrees of freedom of test statistics, if known. Otherwise assumed
to be from N(0, 1).}

\item{methods}{vector of methods to be used. Default c('locfdr','fdrtool','qvalue')}

\item{asym}{boolean, whether to consider an asymmetric simulation model.}

\item{focus_metric}{which metric to prioritize in the grid search.
Must be one of c('pr','roc','brier','Fdrerror')}

\item{large_abs_metric}{if TRUE, only consider focus_metric looking at the
large absolute value test statistics (specifically, top quartile of abs(t))}

\item{parallel}{if TRUE, processes are run in parallel}

\item{verbose}{if TRUE, status updates will be displayed}
}
\value{
\itemize{
\item fdr local false discovery rates
\item Fdr left tail false discovery rates
\item pi0 proportion of null test statistics
\item top_grid dataframe containing hyperparameter sets
that were ensembled over and their metrics on simulated daa
\item all_grids dataframe containing all hyperparameter sets considered
and their metrics on simulated data
\item fit parameters used for simulation step
}
}
\description{
use ensemble methods to estimate local (fdr) and tail-end (Fdr)
false discovery rates.
}
