% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridsemblefdr.R
\name{gridsemble}
\alias{gridsemble}
\title{Gridsemble}
\usage{
gridsemble(
  test_statistics,
  df = NULL,
  to_pval_function = function(test_statistics) {
     p_from_t(test_statistics, df = df)

    },
  ensemble_size = 10,
  n_synthetic = 10,
  locfdr_grid = "default",
  fdrtool_grid = "default",
  qvalue_grid = "default",
  synthetic_size = length(test_statistics),
  n_workers = max(parallel::detectCores() - 2, 1),
  parallel = min(TRUE, n_workers > 1),
  verbose = TRUE,
  drop_pi0_1 = TRUE
)
}
\arguments{
\item{test_statistics}{vector, test statistics}

\item{df}{integer, degrees of freedom of test statistics null t-distribution,
otherwise assumed standard normal. Ignored if \code{to_pval_function} is provided.}

\item{to_pval_function}{function, converts test statistics vector to a
p-value vector. Default assumes t-distribution with given df under the null.}

\item{ensemble_size}{integer, number of models to ensemble}

\item{n_synthetic}{integer, number of datasets to simulate and perform grid search
over, or 0 for models to be randomly selected.}

\item{locfdr_grid}{data.frame or 'default', rows are possible hyperparameters
for locfdr, or \code{build_locfdr_grid} will be run with default values, NULL to
exclude locfdr from gridsemble}

\item{fdrtool_grid}{data.frame or 'default', rows are possible
hyperparameters for fdrtool, or \code{build_fdrtool_grid} will be run with default
values, NULL to exclude fdrtool from gridsemble}

\item{qvalue_grid}{data.frame or 'default', rows are possible hyperparameters
for qvalue, or \code{build_qvalue_grid} will be run with default values, NULL to
exclude qvalue from gridsemble}

\item{synthetic_size}{integer, size of synthetic datasets. Defaults to
\code{length(test_statistics)}. Can be reduced for time/memory constraints.}

\item{n_workers}{integer, number of cores to use if parallel}

\item{parallel}{boolean, whether to utilize parallelization}

\item{verbose}{boolean, whether to report status}
}
\value{
\itemize{
\item \code{fdr}: vector, estimated local false discovery rates
\item \code{Fdr}: vector, estimated left tail-end false discovery rates
\item \code{pi0}: double, estimated proportion of null test statistics
\item \code{top_grid}: data.frame, hyperparameter sets in ensemble and
their estimated metrics on each simulated dataset
\item \code{all_grids}: data.frame, all hyperparameter sets considered
and their estimated metrics on simulated data
\item \code{working_model}: list, parameters of working model
}
}
\description{
use ensemble methods to estimate local (fdr) and tail-end (Fdr)
false discovery rates.
}
\examples{
set.seed(123)
test_statistics = c(rnorm(800), runif(100, -10, -5), runif(100, 5, 10))
res = gridsemble(test_statistics)
res$pi0
res$fdr
}
