% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridsemblefdr.R
\name{gridsemble}
\alias{gridsemble}
\title{Gridsemble}
\usage{
gridsemble(
  test_statistics,
  df = NULL,
  locfdr_grid = build_locfdr_grid(test_statistics, lower_pi0 = lower_pi0, parallel =
    parallel, n_workers = n_workers, verbose = verbose),
  fdrtool_grid = build_fdrtool_grid(test_statistics, lower_pi0 = lower_pi0, parallel =
    parallel, n_workers = n_workers, verbose = verbose),
  qvalue_grid = build_qvalue_grid(test_statistics, df = df, lower_pi0 = lower_pi0,
    parallel = parallel, n_workers = n_workers, verbose = verbose),
  nsim = 10,
  ensemble_size = 10,
  lower_pi0 = 0.7,
  sim_size = length(test_statistics),
  parallel = TRUE,
  n_workers = max(parallel::detectCores() - 2, 1),
  parallel_param = NULL,
  verbose = TRUE
)
}
\arguments{
\item{test_statistics}{vector, test statistics}

\item{df}{integer, degrees of freedom of test statistics, if known. Otherwise assumed
to be from N(0, 1).}

\item{locfdr_grid}{data.frame, each row is a set of hyperparameters for locfdr}

\item{fdrtool_grid}{data.frame, each row is a set of hyperparameters for fdrtool}

\item{qvalue_grid}{data.frame, each row is a set of hyperparameters for qvalue}

\item{nsim}{integer, number of datasets to simulate and grid search over.
If 0, no datasets are simulated and model(s) are randomly selected.}

\item{ensemble_size}{integer, number of models chosen to ensemble over.}

\item{lower_pi0}{double, lower bound cutoff of pi0 for model consideration.
Default 0.7 (assume majority null tests).}

\item{sim_size}{integer, size of simulated datasets, default number of test statistics}

\item{parallel}{boolean, whether to utilize parallelization}

\item{n_workers}{integer, number of cores to use if parallel, default 2 less than available.}

\item{parallel_param}{BiocParallel object, specified to run in parallel or NULL}

\item{verbose}{boolean}
}
\value{
\itemize{
\item fdr local false discovery rates
\item Fdr left tail false discovery rates
\item pi0 proportion of null test statistics
\item top_grid dataframe containing hyperparameter sets
that were ensembled over and their metrics on simulated daa
\item all_grids dataframe containing all hyperparameter sets considered
and their metrics on simulated data
\item generating_model parameters used for simulation step
}
}
\description{
use ensemble methods to estimate local (fdr) and tail-end (Fdr)
false discovery rates.
}
