% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metrics}
\alias{metrics}
\title{Compute all metrics}
\usage{
metrics(fdr, Fdr, truth, true_Fdr, topq)
}
\arguments{
\item{fdr}{estimated fdr values}

\item{Fdr}{estimated (tail end) Fdr values}

\item{truth}{truth values, vector of 0 and 1s}

\item{true_Fdr}{true Fdr values}

\item{topq}{vector of booleans for whether each test statistic
is in the largest absolute quantile}
}
\value{
labeled list of metrics
}
\description{
Compute all metrics
}
