% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metrics}
\alias{metrics}
\title{Compute all metrics}
\usage{
metrics(fdr, true_fdr, topq)
}
\arguments{
\item{fdr}{vector, estimated fdr values}

\item{true_fdr}{vector, true fdr values}

\item{topq}{vector, booleans for whether each test statistic
is in the largest absolute quantile}
}
\value{
labeled list of metrics
}
\description{
Compute all metrics
}
