% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_sim.R
\name{mix}
\alias{mix}
\title{Mixture pdf}
\usage{
mix(t, pi0, sigmasq0, sigmasq1)
}
\arguments{
\item{t}{double, test statistic}

\item{pi0}{double, proportion null}

\item{sigmasq0}{double, variance of null distribution}

\item{type}{string, one of c('symmetric')}
}
\value{
value of mixture density pdf at value t
}
\description{
Mixture pdf
}
