% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_sim.R
\name{null}
\alias{null}
\title{Null density}
\usage{
null(t, sigma_0)
}
\arguments{
\item{t}{test statistic}

\item{sigma_0}{variance of null distribution}
}
\value{
value of null density pdf at value t
}
\description{
Null density
}
