% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generating_model.R
\name{null}
\alias{null}
\title{Null pdf}
\usage{
null(t, sigmasq0)
}
\arguments{
\item{t}{double, test statistic}

\item{sigmasq0}{double, variance of null density}
}
\value{
double, value of null density pdf at value t
}
\description{
Null pdf
}
