% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_qvalue_row.R
\name{p_from_t}
\alias{p_from_t}
\title{P from T}
\usage{
p_from_t(t, df = NULL, sides = "one")
}
\arguments{
\item{t}{vector of t statistics}

\item{df}{degrees of freedom to compute p-value from test statistics,
assume standard normal if NULL}
}
\value{
vector of p-values
}
\description{
get p-value from test statistic
}
