% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{parlapply}
\alias{parlapply}
\title{general lapply function with specification of parallel or not}
\usage{
parlapply(X, FUN, parallel_param, ...)
}
\arguments{
\item{X}{a vector (atomic or list) or an expression object.}

\item{FUN}{the function to be applied to each element of X}

\item{parallel_param}{BiocParallel object, specified to run in parallel or NULL}

\item{...}{additional parameters to \code{BiocParallel::bplapply} or \code{lapply}}
}
\value{
result of lapply
}
\description{
general lapply function with specification of parallel or not
}
