% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_fdrtool_grid.R
\name{reduce_fdrtool_grid}
\alias{reduce_fdrtool_grid}
\title{Reduce fdrtool grid}
\usage{
reduce_fdrtool_grid(test_statistics, fdrtool_grid, verbose = FALSE)
}
\arguments{
\item{test_statistics}{vector of test statistics}

\item{fdrtool_grid}{data frame where each row is a possible set of hyperparameters for fdrtool}

\item{verbose}{}
}
\value{
dataframe where each row is a possible set of hyperparameters for fdrtool
}
\description{
Reduces grid to parameter combinations
that can run fdrtool on the provided data without error
}
