% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_locfdr_grid.R
\name{reduce_locfdr_grid}
\alias{reduce_locfdr_grid}
\title{Reduce locfdr grid}
\usage{
reduce_locfdr_grid(t, locfdr_grid, verbose = FALSE)
}
\arguments{
\item{t}{vector of test statistics}

\item{locfdr_grid}{data frame where each}

\item{verbose}{}
}
\value{
dataframe where each row is a possible set of hyperparameters for locfdr
}
\description{
Reduces grid to parameter combinations
that can run on the provided data without error
}
