% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_locfdr_grid.R
\name{reduce_locfdr_grid}
\alias{reduce_locfdr_grid}
\title{Reduce locfdr grid}
\usage{
reduce_locfdr_grid(
  test_statistics,
  locfdr_grid,
  lower_pi0,
  parallel_param = NULL,
  verbose = FALSE
)
}
\arguments{
\item{test_statistics}{vector, test statistics}

\item{locfdr_grid}{data.frame, each row is a possible set of hyperparameters for locfdr}

\item{lower_pi0}{double, exclude hyperparameter combinations that give pi0 estimates below this threshold}

\item{parallel_param}{BiocParallel object, specified to run in parallel or NULL}

\item{verbose}{boolean}
}
\value{
data.frame, each row is a possible set of hyperparameters for locfdr
}
\description{
Reduces grid to parameter combinations
that can run locfdr on the provided data without error
}
