% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_qvalue_grid.R
\name{reduce_qvalue_grid}
\alias{reduce_qvalue_grid}
\title{Reduce qvalue grid}
\usage{
reduce_qvalue_grid(
  test_statistics,
  qvalue_grid,
  df = NULL,
  parallel = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{test_statistics}{vector of test statistics}

\item{qvalue_grid}{data frame where each row is a possible set of hyperparameters for qvalue}

\item{verbose}{}
}
\value{
dataframe where each row is a possible set of hyperparameters for qvalue
}
\description{
Reduces grid to parameter combinations
that can run qvalue on the provided data without error
}
