% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_qvalue_grid.R
\name{reduce_qvalue_grid}
\alias{reduce_qvalue_grid}
\title{Reduce qvalue grid}
\usage{
reduce_qvalue_grid(
  test_statistics,
  qvalue_grid,
  lower_pi0,
  df = NULL,
  parallel_param = NULL,
  verbose = FALSE
)
}
\arguments{
\item{test_statistics}{vector, test statistics}

\item{qvalue_grid}{data.frame, each row is a possible set of hyperparameters for qvalue}

\item{lower_pi0}{double, exclude hyperparameter combinations that give pi0 estimates below this threshold}

\item{df}{integer, degrees of freedom of test statistics, if known}

\item{parallel_param}{BiocParallel object, specified to run in parallel or NULL}

\item{verbose}{boolean}
}
\value{
data.frame, each row is a possible set of hyperparameters for qvalue
}
\description{
Reduces grid to parameter combinations
that can run qvalue on the provided data without error
}
