% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_fdrtool_row.R
\name{run_fdrtool_row}
\alias{run_fdrtool_row}
\title{Run fdrtool}
\usage{
run_fdrtool_row(t, fdrtool_grid, row)
}
\arguments{
\item{t}{vector of test statistics}

\item{fdrtool_grid}{data frame where each row is a set of hyperparameters}

\item{row}{row of fdrtool_grid, i.e. which set of hyperparameters to run fdrtool with}
}
\value{
\itemize{
\item fdr - estimated local false discovery rates
\item Fdr - estimated left tail-end false discovery rates
\item pi0 - estimated proportion of tests that are null
}
}
\description{
Run fdrtool with a specific set of parameters
}
\examples{

}
