% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_locfdr_row.R
\name{run_locfdr_row}
\alias{run_locfdr_row}
\title{Run locfdr}
\usage{
run_locfdr_row(test_statistics, locfdr_grid, row, returnFdr = TRUE)
}
\arguments{
\item{test_statistics}{vector, test statistics}

\item{locfdr_grid}{data.frame, each row is a set of hyperparameters}

\item{row}{integer, row of locfdr_grid}

\item{returnFdr}{boolean, whether to calculate Fdr form fdr}
}
\value{
\itemize{
\item fdr - estimated local false discovery rates
\item Fdr - estimated left tail-end false discovery rates if returnFdr = TRUE
\item pi0 - estimated proportion of tests that are null
}
}
\description{
Run locfdr with a specific set of parameters
}
