% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_qvalue_row.R
\name{run_qvalue_row}
\alias{run_qvalue_row}
\title{Run qvalue}
\usage{
run_qvalue_row(
  test_statistics,
  qvalue_grid,
  row,
  df = NULL,
  returnFdr = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{test_statistics}{vector, test statistics}

\item{qvalue_grid}{data.frame, each row is a set of hyperparameters}

\item{row}{integer, row of qvalue_grid}

\item{df}{integer, degrees of freedom to compute p-value from test statistics,
assume standard normal if NULL}

\item{verbose}{boolean}
}
\value{
\itemize{
\item fdr - estimated local false discovery rates
\item Fdr - estimated left tail-end false discovery rates
\item pi0 - estimated proportion of tests that are null
}
}
\description{
Run qvalue with a specific set of parameters
}
