% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_sim.R
\name{sample_alternative}
\alias{sample_alternative}
\title{Sample from alternative pdf using metropolis hastings}
\usage{
sample_alternative(N, sigmasq1, burn_in = 1000, t_init = 1)
}
\arguments{
\item{N}{number of test statistics to sample}

\item{sigmasq1}{variance of alternative distribution}

\item{burn_in}{number of iterations to use as burn in and discard}
}
\value{
test_statistics vector of test statistics
}
\description{
Sample from alternative pdf using metropolis hastings
}
