% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generating_model.R
\name{sample_alternative}
\alias{sample_alternative}
\title{Sample from alternative pdf using metropolis hastings}
\usage{
sample_alternative(N, sigmasq1, burn_in = 1000, t_init = 1)
}
\arguments{
\item{N}{integer, number of test statistics to sample}

\item{sigmasq1}{double, variance of alternative distribution}

\item{burn_in}{integer, number of iterations to use as burn in and discard}

\item{t_init}{double, starting point for metropolis hastings}
}
\value{
vector, test statistics
}
\description{
Sample from alternative pdf using metropolis hastings
}
