% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generating_model.R
\name{sample_null}
\alias{sample_null}
\title{Sample from null density}
\usage{
sample_null(n, sigmasq0)
}
\arguments{
\item{n}{integer, sample size}

\item{sigmasq0}{double, variance of null distribution}
}
\value{
vector, test statistics sampled from null density N(0, sigmasq0)
}
\description{
Sample from null density
}
