% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_sim.R
\name{simulate_from_fit}
\alias{simulate_from_fit}
\title{Simulate from fit}
\usage{
simulate_from_fit(n, fit)
}
\arguments{
\item{n}{sample size}

\item{fit}{result of fit_sim()}
}
\value{
data frame of n simulated t-statistic truth pairs
}
\description{
simulate a dataset of size n from the mixture and record truth label
of each value (null vs alternative)
}
